requirements.txt
boto3>=1.26.0
pandas>=2.0.0
scikit-learn>=1.3.0
numpy>=1.24.0
pyarrow>=12.0.0
statsmodels>=0.14.0
matplotlib>=3.7.0
seaborn>=0.12.0
pyyaml>=6.0
joblib>=1.3.0
python-dateutil>=2.8.2
scipy>=1.11.0
